Fix file backup in dev mode & fatal restore error
When we were restoring the engine from a previous backup where there was
no active DWH, Cinderlib installation or Keycloak database, the restore
always resulted in a FATAL error:

-----
Rewriting /home/build/ovirt//etc/ovirt-engine/engine.conf.d/10-setup-database.conf
/home/build/ovirt/share/ovirt-engine/bin/engine-backup.sh: line 1419: [: syntax error: `-' unexpected
FATAL: Can not find /home/build/ovirt//etc/ovirt-engine/engine.conf.d/10-setup-cinderlib-database.conf
-----

This is because of a check that is done when restoring the engine. When
only restoring the DB, the value ${CHANGE_DB_CREDENTIALS} is true and
because of the OR condition, the false flag of for example
${CHANGE_DWH_CREDENTIALS} does not have any effect anymore. By removing
the CHANGE_DB_CREDENTIALS variable, the condition is dependent on only
the flag of the type of DB itself and the DB_USER.

There were also issues with running a backup in development mode.
When installing a development environment, a path is specified with the
PREFIX value. However, the backup script does not change the paths that
it is backing up relative to this PREFIX value. This has also been
changed to that it does backup the correct paths.

Signed-off-by: Brooklyn Dewolf <contact@brooklyn.gent>